Added operators, syntax changes, more built-in methods
Added the unary increment and decrement operators, as well as the power operator. Also changes var -> assign, fun -> function, or -> ||, and -> &&. Added sin, cos, cot, and tan methods, as well as toRad for converting from degrees to radians, and a type method that returns the type of a parameter in the form of a string to allow for basic control of types.